no need to cast patch->columnofs to uint16_t

+++ i_amiv16.c
@@ -1,7 +1,7 @@
 /*-----------------------------------------------------------------------------
  *
  *
- *  Copyright (C) 2025 Frenkel Smeijers
+ *  Copyright (C) 2025-2026 Frenkel Smeijers
  *
  *  This program is free software; you can redistribute it and/or
  *  modify it under the terms of the GNU General Public License
@@ -781,7 +781,7 @@ void V_DrawPatchScaled(int16_t x, int16_t y, const patch_t __far* patch)
 		else if (dc_x >= SCREENWIDTH)
 			break;
 
-		const column_t *column = (const column_t *)((const byte *)patch + (uint16_t)patch->columnofs[col >> 8]);
+		const column_t *column = (const column_t *)((const byte *)patch + patch->columnofs[col >> 8]);
 
 		x = dc_x & 7;
 		uint8_t andmask = ~(0x80 >> x);
